cast to int

+++ pixelate.py
@@ -46,15 +46,15 @@ print(math.floor(image.size[1]/pixelSize))
 
 image = image.resize(
     (
-        math.floor(image.size[0]/pixelSize),
-        math.floor(image.size[1]/pixelSize)
+        int(math.floor(image.size[0]/pixelSize)),
+        int(math.floor(image.size[1]/pixelSize))
     ),
     Image.NEAREST
 )
 image = image.resize(
     (
-      math.floor(image.size[0]*pixelSize),
-      math.floor(image.size[1]*pixelSize)
+      int(math.floor(image.size[0]*pixelSize)),
+      int(math.floor(image.size[1]*pixelSize))
     ),
     Image.NEAREST
 )
@@ -64,15 +64,6 @@ image.thumbnail(maxsize, Image.NEAREST)
 image.save('output.jpg')
 
 sense = SenseHat()
-
-# sense.set_pixel(2, 2, (0, 0, 255))
-# sense.set_pixel(4, 2, (0, 0, 255))
-# sense.set_pixel(3, 4, (100, 0, 0))
-# sense.set_pixel(1, 5, (255, 0, 0))
-# sense.set_pixel(2, 6, (255, 0, 0))
-# sense.set_pixel(3, 6, (255, 0, 0))
-# sense.set_pixel(4, 6, (255, 0, 0))
-# sense.set_pixel(5, 5, (255, 0, 0))
 
 
 def test():
